src/api/serializers.py
from rest_framework import serializers
from src.api.models import Project


class ProjectSerializer(serializers.ModelSerializer):
    class Meta:
        model = Project
        fields = ('id', 'name', 'url', 'port')
