requirements.txt
certifi==2017.11.5
chardet==3.0.4
Django==2.0.1
djangorestframework==3.7.7
docker==2.7.0
docker-pycreds==0.2.1
idna==2.6
pytz==2017.3
requests==2.18.4
six==1.11.0
urllib3==1.22
websocket-client==0.46.0
